removed orphaned games from map
