+++ frontend/mastermind-frontend/app/routes/login.tsx
@@ -26,6 +26,7 @@ const LoginForm: React.FC = () => {
             const data = await response.json();
             if (response.ok) {
                 setIsLoggedIn(true)
+
                 navigate('/home');
             } else {
                 setError(data.message);
