+++ sessionInfo.md
@@ -1,0 +1,9 @@
+lerDecorator : New StandardWebSocketSession[id=fe0f5f7e-5982-6c00-bd2d-0da2aea05ef6, uri=ws://localhost:8080/ws]
+2025-09-05T00:50:19.453-07:00 DEBUG 92005 --- [mastermind] [nboundChannel-2] org.springframework.web.SimpLogging      : Processing CONNECT user=MrGuy session=fe0f5f7e-5982-6c00-bd2d-0da2aea05ef6
+2025-09-05T00:50:19.460-07:00 DEBUG 92005 --- [mastermind] [nboundChannel-5] org.springframework.web.SimpLogging      : Processing SUBSCRIBE /topic/mp id=sub-0 session=fe0f5f7e-5982-6c00-bd2d-0da2aea05ef6
+2025-09-05T00:50:19.462-07:00 DEBUG 92005 --- [mastermind] [nio-8080-exec-2] o.s.web.servlet.DispatcherServlet        : GET "/ws", parameters={}
+2025-09-05T00:50:19.463-07:00 DEBUG 92005 --- [mastermind] [nio-8080-exec-2] o.s.w.s.s.s.WebSocketHandlerMapping      : Mapped to org.springframework.web.socket.server.support.WebSocketHttpRequestHandler@4a842e8a
+2025-09-05T00:50:19.463-07:00 DEBUG 92005 --- [mastermind] [nio-8080-exec-2] o.s.w.s.s.s.WebSocketHttpRequestHandler  : GET /ws
+2025-09-05T00:50:19.464-07:00 DEBUG 92005 --- [mastermind] [nio-8080-exec-2] o.s.web.servlet.DispatcherServlet        : Completed 101 SWITCHING_PROTOCOLS
+2025-09-05T00:50:19.465-07:00 DEBUG 92005 --- [mastermind] [nio-8080-exec-2] s.w.s.h.LoggingWebSocketHandlerDecorator : New StandardWebSocketSession[id=e0f09758-ec25-9b0d-77bf-528dcda07be5, uri=ws://localhost:8080/ws]
+2025-09-05T00:50:19.466-07:00 DEBUG 92005 --- [mastermind] [nboundChannel-8] org.springframework.web.SimpLogging      : Processing CONNECT user=usernamename session=e0f09758-ec25-9b0d-77bf-528dcda07be5
