+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=mastermind
 spring.datasource.url=jdbc:postgresql://localhost:5432/mastermind
 spring.datasource.username=nolimit
 spring.datasource.password=
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
